# render 5 initial items in search screen on mobile