--- a/src/components/SelectionList/types.ts
+++ b/src/components/SelectionList/types.ts
@@ -633,6 +633,9 @@ type BaseSelectionListProps<TItem extends ListItem> = Partial<ChildrenProps> & {
 
     /** Called when scrollable content view of the ScrollView changes */
     onContentSizeChange?: (w: number, h: number) => void;
+
+    /** Initial number of items to render */
+    initialNumToRender?: number;
 } & TRightHandSideComponent<TItem>;
 
 type SelectionListHandle = {